added comments, renamed test class and methods
