+++ producer/src/test/kotlin/streams/integrations/KafkaEventRouterEnterpriseTSE.kt
@@ -20,7 +20,7 @@ import streams.utils.StreamsUtils
 import java.util.concurrent.TimeUnit
 import kotlin.test.assertEquals
 import org.neo4j.function.ThrowingSupplier
-import streams.integrations.CompactionStrategyTestCommon.assertTopicFilled
+import streams.integrations.KafkaLogCompactionTestCommon.assertTopicFilled
 
 class KafkaEventRouterEnterpriseTSE {
